pyproject.toml
[project]
name = "blog-generation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.74",
    "langchain-groq>=0.3.7",
    "langgraph>=0.6.4",
    "langgraph-cli[inmen]>=0.3.6",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
    "watchdog>=6.0.0",
]
